move files to migration folder
