+++ module/services/_module.mjs
@@ -1,4 +1,4 @@
 export {default as RollProcessor} from "./roll-processor.mjs";
 export {default as JournalService} from "./journal-service.mjs";
-export {default as MigrationManager} from "./migration-manager.mjs";
-export {default as TypeTransformationManager} from "./type-transformation-manager.mjs";
+export {default as MigrationManager} from "./migrations/migration-manager.mjs";
+export {default as TypeTransformationManager} from "./migrations/type-transformation-manager.mjs";
